app/requirements.txt
djangorestframework==3.14.0
requests
celery[redis]
psycopg2-binary
redis